% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_origin.R
\name{clonal_origin}
\alias{clonal_origin}
\title{ClonalOrigin.}
\usage{
clonal_origin(
  tree_filename,
  data_filename,
  output_filename,
  w = NA,
  x = NA,
  y = NA,
  z = NA,
  T = NA,
  R = NA,
  D = NA,
  s = NA,
  S = NA,
  r = NA,
  t = NA,
  U = NA,
  G = NA,
  a = NA,
  m = NA,
  i = NA,
  f = NA,
  v = NA,
  h = NA,
  V = NA
)
}
\arguments{
\item{tree_filename}{The file containing the clonal frame tree.}

\item{data_filename}{The file containing the aligned sequence data.}

\item{output_filename}{The file to which the output will be written.}

\item{w}{Sets the number of pre burn-in iterations (default is 100000).}

\item{x}{Sets the number of burn-in iterations (default is 100000).}

\item{y}{Sets the number of iterations after burn-in (default is 100000).}

\item{z}{Sets the number of iterations between samples (default is 100).}

\item{T}{Sets the value of theta. Write "sNUM" in a string to set a per-site rate of NUM.}

\item{R}{Sets the value of rho. Write "sNUM" in a string to set a per-site rate of NUM.}

\item{D}{Sets the value of delta.}

\item{s}{Use given seed to initiate random number generator.}

\item{S}{Takes a string. If in the format "NUM,SEED", runs on a subset of NUM regions determined by seed SEED; if in the format "NUM/NUM/../NUM", runs on a specified region(s) given by each NUM.}

\item{r}{Perform r tempered steps between topological updates (default:0).}

\item{t}{Tempered at "temperature" t for topological updates (default:1).}

\item{U}{If TRUE, Start from UPGMA tree, rather than the default random tree.}

\item{G}{Greedily compute the "best fit" tree, given the recombination observed on the current tree.  If the input is negative and a previous run is provided, the tree is calculated from all observed values. If the input is positive, a "greedy move" is performed with this weight (see argument a). Note that this is NOT an MCMC move and causes bias.}

\item{a}{A vector or comma separated string setting the ELEVEN (real valued) move weightings to the given vector. The weightings need not sum to 1, but must be in the following order: MoveRho (ignored if not needed); MoveDelta (ignored if not needed); MoveTheta (ignored if not needed); MoveRemEdge; MoveAddEdge; MoveSiteChange; MoveTimeChange; MoveEdgeChange; MoveAgeClonal; MoveScaleTree; MoveRegraftClonal.}

\item{m}{A vector or comma separated string setting THREE numbers required for MAIS implementation: gamma Power used in the annealing computation, gamma>1 means steps for jumping up concentrate near start, gamma=1 means steps are equidistant; T Number of steps for AIS, T=1 corresponds to no intermediate steps; N Number of replications for MAIS, instead of using N=1 (which corresponds to pure AIS-RJMCMC) use N=0.}

\item{i}{A vector or comma separated string setting the SIX parameters for creating random Recombination Trees under the inference model. The parameters are: N    (integer)    The number of sequences in the sample (default 10); n_B    (integer)    The number of block boundaries in the sample (default 8); l_B    (integer)    The length of each block: L=n_B * l_B (default 500); delta (real) The average length of imports (default 500.0); theta (real) The mutation rate NOT per site (default 100.0); rho    (real) The recombination rate NOT per site (default 50.0).}

\item{f}{If TRUE, forbid topology changes, (allowing updates of coalescence times).}

\item{v}{If TRUE, use verbose mode.}

\item{h}{If TRUE, display a help message.}

\item{V}{If TRUE, print ClonalOrigin Version info.}
}
\value{
Writes output to a file specified by output_filename.
}
\description{
ClonalOrigin.
}
