% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RClonOr-package.R
\docType{package}
\name{RClonOr}
\alias{RClonOr}
\title{RClonOr: ClonalOrigin in R.}
\description{
A version of https://github.com/fmedina7/ClonOr_cpp that is callable from R.
}
\section{Functions}{

Some functions.
}

